Desktop/MenuScraper_Playwright/MenuScraper_Playwright/requirements_ml.txt
# ML-Enhanced Menu Scraper Requirements
# Core dependencies for machine learning functionality

# Existing scraper dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10

# Machine Learning and NLP
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Additional ML utilities
huggingface-hub>=0.17.0
tokenizers>=0.14.0
safetensors>=0.4.0

# Text processing
nltk>=3.8.1
spacy>=3.7.0

# Data handling
jsonschema>=4.19.0
pyyaml>=6.0.1

# Performance and monitoring
tqdm>=4.66.0
psutil>=5.9.0

# Optional: Advanced ML features
# tensorflow>=2.13.0  # Uncomment if using TensorFlow models
# opencv-python>=4.8.0  # Uncomment for advanced image processing
# selenium>=4.15.0  # Uncomment for additional browser automation

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0